+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Library: Added immutable client-lifetime traffic snapshots through `trafficStats()`.
+
 ## [3.1.0] - 2026-07-13
 
 ### BREAKING